+++ diesel/src/pg/types/network_address.rs
@@ -20,11 +20,6 @@ const AF_INET: u8 = libc::AF_INET as u8;
 const PGSQL_AF_INET: u8 = AF_INET;
 const PGSQL_AF_INET6: u8 = AF_INET + 1;
 
-// https://github.com/postgres/postgres/blob/502a3832cc54c7115dacb8a2dae06f0620995ac6/src/include/catalog/pg_type.h#L435-L443
-primitive_impls!(MacAddr -> (pg: (829, 1040)));
-primitive_impls!(Inet -> (pg: (869, 1041)));
-primitive_impls!(Cidr -> (pg: (650, 651)));
-
 #[allow(dead_code)]
 mod foreign_derives {
     use super::*;